requirements.txt
langchain
langchain-community
langchain-groq
python-dotenv
pydantic
typing

streamlit
pandas
plotly



